Add support for setting a custom installation path of Android Studio

+++ SecFontsFix.ps1
@@ -1,6 +1,6 @@
 # Function to check if running as administrator
 function Run-as-admin {
-    $currentUser = New-Object Security.Principal.WindowsPrincipal $([Security.Principal.WindowsIdentity]::GetCurrent())
+    $currentUser = New-Object Security.Principal.WindowsPrincipal $([Security.Principal.WindowsIdentity]::GetCurrent() )
     $currentUser.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
 }
 
@@ -14,10 +14,18 @@ if (-not (Run-as-admin)) {
 Write-Host "Running your PowerShell script with elevated privileges..."
 
 function Show-Menu {
+    param (
+        [string]$CurrentInstallDir
+    )
     Write-Host "Select an option:"
-    Write-Host "1. Sec Fonts Render Fix"
-    Write-Host "2. Restore Stock Fonts"
-    Write-Host "3. Exit"
+    if ($CurrentInstallDir) {
+        Write-Host "1. Set Android Studio's Installation Directory (Current: $CurrentInstallDir)"
+    } else {
+        Write-Host "1. Set Android Studio's Installation Directory (No directory set)"
+    }
+    Write-Host "2. Sec Fonts Render Fix"
+    Write-Host "3. Restore Stock Fonts"
+    Write-Host "4. Exit"
 }
 
 # Function to merge XML contents
@@ -39,22 +47,67 @@ function Merge-XmlContents {
     }
 }
 
+# Path to config file
+$configFile = Join-Path $PSScriptRoot "config.txt"
+
+# Function to get or set install directory
+function Get-InstallDirectory {
+    param (
+        [string]$ConfigFilePath
+    )
+    if (Test-Path $ConfigFilePath) {
+        $dir = Get-Content $ConfigFilePath -ErrorAction SilentlyContinue | Select-Object -First 1
+        if ($dir -and (Test-Path $dir)) {
+            return $dir
+        }
+    }
+    return $null
+}
+
+function Set-InstallDirectory {
+    param (
+        [string]$ConfigFilePath
+    )
+    Write-Host "Enter the  installation directory path of android studio. E.g: C:\Program Files\Android\Android Studio"
+    $userDir = Read-Host "Hit Enter key when done or to go back to the main menu"
+    $userDir = $userDir.Trim()
+    if ([string]::IsNullOrWhiteSpace($userDir)) {
+        Write-Host "No directory entered."
+        Start-Sleep -Seconds 2
+        return $null
+    }
+    if (-not (Test-Path $userDir)) {
+        Write-Host "Directory does not exist." -ForegroundColor Red
+        Start-Sleep -Seconds 2
+        return $null
+    } else {
+        $fontsDirectory = Join-Path $userDir "plugins\design-tools\resources\layoutlib\data\fonts"
+        if (-not (Test-Path $fontsDirectory)){
+            Write-Host "Invalid directory; $fontsDirectory not found." -ForegroundColor Red
+            Start-Sleep -Seconds 2
+            return $null
+        }
+    }
+    Set-Content -Path $ConfigFilePath -Value $userDir
+    Write-Host "Installation directory set to: $userDir"
+    return $userDir
+}
+
+$installDir = Get-InstallDirectory -ConfigFilePath $configFile
+
 # Function to run Script 1
 function Run-SecFontsFix {
-	Write-Host "`n`n`n"
+    Write-Host "`n`n"
     Write-Host "Start Sec Fonts Render Fix..."
     $secFontsPath = "$PSScriptRoot\secfonts"
-    Set-Location "$env:ProgramFiles\Android"
-
     $found = $false
 
-    Get-ChildItem -Directory | ForEach-Object {
-        $dir = $_.FullName
-        $fontsDirectory = "$dir\plugins\design-tools\resources\layoutlib\data\fonts"
+    if (Test-Path $installDir) {
+        $fontsDirectory = Join-Path $installDir "plugins\design-tools\resources\layoutlib\data\fonts"
         if (Test-Path $fontsDirectory) {
-			Write-Host "Working on $fontsDirectory..."
+            Write-Host "Working on $fontsDirectory..."
             $found = $true
-            $fontsBackupDirectory = "$dir\plugins\design-tools\resources\layoutlib\data\fonts_backup"
+            $fontsBackupDirectory = Join-Path $installDir "plugins\design-tools\resources\layoutlib\data\fonts_backup"
             if (!(Test-Path $fontsBackupDirectory)) {
                 Write-Host "Creating backup of original fonts..."
                 Copy-Item -Path $fontsDirectory -Destination $fontsBackupDirectory -Recurse
@@ -65,46 +118,45 @@ function Run-SecFontsFix {
             robocopy $secFontsPath $fontsDirectory *.ttf /IS /IM /XC /XN /njh /njs /ndl /nc /ns /np /nfl /ndl
             Write-Host "... done!"
             Write-Host ""
-			# Merge contents of fonts.xml with patch to AS fonts.xml
+            # Merge contents of fonts.xml with patch to AS fonts.xml
             $fontsXmlPath = Join-Path $fontsDirectory "fonts.xml"
-			if (Test-Path $fontsXmlPath) {
-				Write-Host "Merging sec fonts.xml to android studio fonts.xml..."
-				$patchFilePath = Join-Path $secFontsPath "fonts.xml"
-				$mainXml = [xml](Get-Content $fontsXmlPath)
-				$patchXml = [xml](Get-Content $patchFilePath)
-				Merge-XmlContents -mainXml $mainXml -patchXml $patchXml
-				$mainXml.Save($fontsXmlPath)
-				Write-Host "... done!"
-			} else {
-				Write-Host "fonts.xml not found in android studio fonts directory!"
-			}
-			Write-Host ""
-			
-			# Merge contents of fonts.xml with patch to AS font_fallback.xml
-            
-            $fontFallbackXmlPath = Join-Path $fontsDirectory "font_fallback.xml"
-			if (Test-Path $fontFallbackXmlPath) {
-				Write-Host "Merging sec fonts.xml to android font_fallback.xml..."
-				$patchFilePath = Join-Path $secFontsPath "fonts.xml"
-				$mainXmlFallback = [xml](Get-Content $fontFallbackXmlPath)
-				$patchXml = [xml](Get-Content $patchFilePath)
-				Merge-XmlContents -mainXml $mainXmlFallback -patchXml $patchXml
-				$mainXmlFallback.Save($fontFallbackXmlPath)
-				Write-Host "... done!"
-			} else {
-				Write-Host "font_fallback.xml not found in android studio fonts directory!"
-			}
-        } else {
-            Write-Host "plugins\design-tools\resources\layoutlib\data\fonts not found in $dir, skipping..."
-        }
-		Write-Host "`n`n`n"
-    }
+            if (Test-Path $fontsXmlPath) {
+                Write-Host "Merging sec fonts.xml to android studio fonts.xml..."
+                $patchFilePath = Join-Path $secFontsPath "fonts.xml"
+                $mainXml = [xml](Get-Content $fontsXmlPath)
+                $patchXml = [xml](Get-Content $patchFilePath)
+                Merge-XmlContents -mainXml $mainXml -patchXml $patchXml
+                $mainXml.Save($fontsXmlPath)
+                Write-Host "... done!"
+            } else {
+                Write-Host "fonts.xml not found in android studio fonts directory!"
+            }
+            Write-Host ""
 
-    if (-not $found) {
-        Write-Host "Error: No plugins\design-tools\resources\layoutlib\data\fonts directory found in any subfolder of ProgramFiles\Android." -ForegroundColor Red
+            # Merge contents of fonts.xml with patch to AS font_fallback.xml
+
+            $fontFallbackXmlPath = Join-Path $fontsDirectory "font_fallback.xml"
+            if (Test-Path $fontFallbackXmlPath) {
+                Write-Host "Merging sec fonts.xml to android font_fallback.xml..."
+                $patchFilePath = Join-Path $secFontsPath "fonts.xml"
+                $mainXmlFallback = [xml](Get-Content $fontFallbackXmlPath)
+                $patchXml = [xml](Get-Content $patchFilePath)
+                Merge-XmlContents -mainXml $mainXmlFallback -patchXml $patchXml
+                $mainXmlFallback.Save($fontFallbackXmlPath)
+                Write-Host "... done!"
+            } else {
+                Write-Host "font_fallback.xml not found in android studio fonts directory!"
+            }
+        }
+
+        if (-not $found) {
+            Write-Host "$fontsDirectory not found." -ForegroundColor Red
+        } else {
+            Write-Host "Done!"
+            Write-Host "Restart android studio to take effect."
+        }
     } else {
-        Write-Host "Done!"
-		Write-Host "Restart android studio to take effect."
+        Write-Host "$installDir does not exist." -ForegroundColor Red
     }
 
     Pause
@@ -112,43 +164,52 @@ function Run-SecFontsFix {
 
 # Function to run Script 2
 function Run-RestoreStockFonts {
-    Set-Location "$env:ProgramFiles\Android"
-    Get-ChildItem -Directory | ForEach-Object {
-        $dir = $_.FullName
-        if (Test-Path "$dir\plugins\design-tools\resources\layoutlib\data\fonts_backup") {
-            Write-Host "Restoring original fonts for $dir from backup..."
-            Remove-Item -Path "$dir\plugins\design-tools\resources\layoutlib\data\fonts" -Recurse -Force
-            Rename-Item "$dir\plugins\design-tools\resources\layoutlib\data\fonts_backup" fonts
+    if (Test-Path $installDir) {
+        $backupFontsDir = Join-Path $installDir "plugins\design-tools\resources\layoutlib\data\fonts_backup"
+        $fontsDir = Join-Path $installDir "plugins\design-tools\resources\layoutlib\data\fonts"
+        if (Test-Path $backupFontsDir) {
+            Write-Host "Restoring $installDir original fonts from backup..."
+            Remove-Item -Path "$fontsDir" -Recurse -Force
+            Rename-Item "$backupFontsDir" fonts
             Write-Host "Done!"
-			Write-Host ""
+            Write-Host ""
         } else {
-            Write-Host "Error: fonts_backup folder not found for $dir!"
+            if (Test-Path $fontsDir) {
+                Write-Host "$backupFontsDir not found."
+            } else {
+                Write-Host "$installDir directory is invalid." -ForegroundColor Red
+            }
         }
+    } else {
+        Write-Host "$installDir does not exist." -ForegroundColor Red
     }
 
     Pause
 }
 
-# Main script loop
+# Main menu loop
 $continue = $true
 while ($continue) {
     Clear-Host
-    Show-Menu
-    $choice = Read-Host "Enter your choice"
+    Show-Menu -CurrentInstallDir $installDir
+    $choice = Read-Host "Enter your choice (1-4)"
     switch ($choice) {
-        '1' {
+        "1" {
+            $installDir = Set-InstallDirectory -ConfigFilePath $configFile
+        }
+        "2" {
             Run-SecFontsFix
         }
-        '2' {
+        "3" {
             Run-RestoreStockFonts
         }
-        '3' {
+        "4" {
             Write-Host "Exiting..."
             sleep 1
             $continue = $false
         }
         default {
-            Write-Host "Invalid choice. Please select again."
+            Write-Host "Invalid selection. Please choose 1-4."
             Start-Sleep -Seconds 1
         }
     }
